requirements.txt
python-dotenv
openai
pyupbit
pyjwt
pandas==1.3.3
pandas_ta==0.3.14b0
schedule
streamlit
selenium
numpy==1.21.4